refactoring: refactor Item/UserItem entities and repos; update purchase logic for quantity increment
