--- src/main/java/efub/cpbr/crumble/item/dto/ItemCountResponse.java
@@ -1,6 +1,0 @@
-package efub.cpbr.crumble.item.dto;
-
-import efub.cpbr.crumble.item.entity.ItemType;
-
-public record ItemCountResponse(ItemType type, int quantity) {
-}
