+++ src/main/java/efub/cpbr/crumble/shop/item/controller/FortuneController.java
@@ -1,8 +1,8 @@
-package efub.cpbr.crumble.item.controller;
+package efub.cpbr.crumble.shop.item.controller;
 
-import efub.cpbr.crumble.item.dto.FortuneAnswerResponse;
-import efub.cpbr.crumble.item.dto.FortuneUseCheckResponse;
-import efub.cpbr.crumble.item.service.FortuneService;
+import efub.cpbr.crumble.shop.item.dto.FortuneAnswerResponse;
+import efub.cpbr.crumble.shop.item.dto.FortuneUseCheckResponse;
+import efub.cpbr.crumble.shop.item.service.FortuneService;
 import efub.cpbr.crumble.user.entity.User;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -12,7 +12,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Tag(name = "Fortune Cookie", description = "포춘쿠키 관련 API")
