+++ src/main/java/efub/cpbr/crumble/shop/item/dto/ItemCountResponse.java
@@ -1,0 +1,6 @@
+package efub.cpbr.crumble.shop.item.dto;
+
+import efub.cpbr.crumble.shop.item.entity.ItemType;
+
+public record ItemCountResponse(ItemType type, int quantity) {
+}
