+++ src/main/java/efub/cpbr/crumble/shop/item/entity/ItemType.java
@@ -1,4 +1,4 @@
-package efub.cpbr.crumble.item.entity;
+package efub.cpbr.crumble.shop.item.entity;
 
 public enum ItemType {
     ERASER,
